Generating .ics calendar file using Jekyll.
`.ics` uses non-inclusive upper bounds on dates, e.g., if the event runs from May 1 to May 2, the end date is shown as May 3 (to support one-day events, presumably).  http://stackoverflow.com/questions/32894339/calculating-with-time-in-jekyll shows the hoops we have to jump through to add one day to a date in Jekyll.  This PR also gets a proper timestamp into the AMY data.