Create 2016-01-12-instructor-debriefing-round-01.html
Here is a brief write up of this morning's debriefing. I'd like to give @k8hertweck a chance to add/make changes as necessary (is the best practices slide show the link you found as well or is there something else?). 

Also, I used the template made by @rgaiacs, but modified it by deleted the section on installation issues (not discusses), changed the "what went wrong" to "what could have gone better", and added an "other comments" section because I wasn't sure where the link to the workshop blog should go.