Ignore Gemfile.lock
Gemfile.lock is generated by bundler when installing dependencies for
this project.

Error from #220 happened because Gemfile.lock was in the repository
and every time the dependencies updated and saved to the *lock file,
the old version from repository overwrote it. And then failed to
build the website.

By ignoring Gemfile.lock we'll be able to generate it on a per-machine
need and it won't be overwritten by building scripts (since it's being
removed from the repository as well).