Fix: intructor markers on the map
There was one marker per airport instead of one marker per instructor.
Now there are multiple markers per airport, each marker corresponds to
one instructor.

Additionally fixed: indentation in `maps.js` (what a nightmare it was;
mix of tabs and spaces, and not equal to 2 or 4 spaces but sometimes
to this strange number of 6).